simplify API by removing loads()/dumps()
It's not so common to need to load/dump from/to a string. But if you
really need to do it, e.g. we use this for tests, it's quite easy to
use a temporary file.
Moreover, we didn't have a way to load/dump from/to a file-like object
that has a write() method. This is far more useful, so we add that here.
This also provides another option for strings via io.StringIO().

Also closes #131 while in the neighbourhood.