explicitly serialise numbers as int or float
This is what the YAML emitter should be doing, but instead
it checks for exact type matches. The JSON emitter is a little
better, as it checks for int and float subclasses, but this
still fails for numpy.int32/int64 types as these are not int
subclasses. They do fit into Python's numeric tower, so we
check for instances of numbers.Integral or numbers.Real.

Closes #156