Fix defaults when using Builder, and add more tests.
The tests here also uncovered a small issue with migration defaults.
If the migration default in question is "demes", then the user must
override to obtain an asymmetric migration. Similarly, if the default is
the "source" and/or "dest", these must be overriden to obtain a
symmetric migration. With the changes here, overrides of 'None' can now
be used with the Builder API.

Closes #214.